+++ tests/unit/datarobot_drum/drum/test_args_parser.py
@@ -49,6 +49,45 @@ def get_args_parser_options(cli_command: List[str]):
     return options
 
 
+@pytest.fixture
+def this_dir():
+    return str(Path(__file__).absolute().parent)
+
+
+@pytest.fixture
+def fit_args(this_dir):
+    return [
+        "fit",
+        "--code-dir",
+        this_dir,
+        "--input",
+        __file__,
+        "--target",
+        "pronounced-tar-ZHAY",
+    ]
+
+
+@pytest.fixture
+def score_args(this_dir):
+    return [
+        "score",
+        "--code-dir",
+        this_dir,
+        "--input",
+        __file__,
+    ]
+
+
+@pytest.fixture
+def server_args(this_dir):
+    return ["server", "--code-dir", this_dir, "--address", "https://allthedice.com"]
+
+
+@pytest.fixture
+def push_args(this_dir):
+    return ["push", "--code-dir", this_dir]
+
+
 class TestDrumHelp:
     @pytest.mark.parametrize(
         "cmd",
@@ -478,40 +517,6 @@ class TestDrApiAccess:
 
 
 class TestUserSecretsArgs:
-    @pytest.fixture
-    def this_dir(self):
-        return str(Path(__file__).absolute().parent)
-
-    @pytest.fixture
-    def fit_args(self, this_dir):
-        return [
-            "fit",
-            "--code-dir",
-            this_dir,
-            "--input",
-            __file__,
-            "--target",
-            "pronounced-tar-ZHAY",
-        ]
-
-    @pytest.fixture
-    def score_args(self, this_dir):
-        return [
-            "score",
-            "--code-dir",
-            this_dir,
-            "--input",
-            __file__,
-        ]
-
-    @pytest.fixture
-    def server_args(self, this_dir):
-        return ["server", "--code-dir", this_dir, "--address", "https://allthedice.com"]
-
-    @pytest.fixture
-    def push_args(self, this_dir):
-        return ["push", "--code-dir", this_dir]
-
     @pytest.fixture(params=["fit_args", "score_args", "server_args", "push_args"])
     def parametrized_args(self, request):
         yield request.getfixturevalue(request.param)
@@ -654,3 +659,16 @@ class TestTritonServerArgs:
                 minimal_score_cmd_args,
                 is_env_variable,
             )
+
+
+class TestRuntimeParametersArgs:
+    @pytest.mark.parametrize("args", ["score_args", "server_args"])
+    def test_runtime_params_valid(self, request, args):
+        args = request.getfixturevalue(args)
+        args.extend([ArgumentsOptions.RUNTIME_PARAMS_FILE, __file__])
+        get_args_parser_options(args)
+
+    def test_runtime_params_invalid(self, fit_args):
+        with pytest.raises(SystemExit), NamedTemporaryFile() as f:
+            fit_args.extend([ArgumentsOptions.RUNTIME_PARAMS_FILE, f.name])
+            get_args_parser_options(fit_args)
